Water energy food nexus multi-objective optimization model
This Jupyter notebook contains the multi-objective optimization (MOO) model that optimizes the water energy food nexus (WEFN) to generate an optimal resource allocation scheme, optimal healthy and sustainable diet composition, in addition to the breakdown of how resources are consumed across sectors and sub-sectors. The model accounts for a variety of nexus interlinkages and factors, the mathematical basis for which are explained in more detail in the following scientific journal paper: "Multi-objective optimization for comprehensive water, energy, food nexus modeling" by Mansour et al. (2023) published in Sustainable Production and Consumption.